Reverted to old code. But working css, login, signup, database corresponds with specific username
