+++ client/login.html
@@ -5,13 +5,17 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
+    <title>Travel-Diary</title>
+    <link rel="stylesheet" href="./login.css">
+    <link href="https://fonts.googleapis.com/css?family=Great+Vibes|Rubik:400i&display=swap" rel="stylesheet">
 </head>
 
 <body>
+    <div class="background"></div>
     <!-- Send to /profile as POST request -->
     <form class="login-form" action="/profile" method="POST">
-        <h1 id=loginHeader>LogIn PAGE</h1>
+        <h1 id=loginHeader>Travel-Diary</h1>
+        <h2 id="please-login">Please Signin</h2>
         <!-- Username -->
         <p>
             <label for="login">Username:</label>
@@ -28,7 +32,8 @@
         </p>
     </form>
     <form class="sign-up-form" action="/signup" method="GET">
-        <input id="sign-up-btn" type="submit" value="Signup">
+        <h3 id="no-account">Don't have an account?</h3>
+        <input id="sign-up-btn" type="submit" value="Signup Here!">
 
     </form>
 </body>
