+++ client/signup.css
@@ -1,0 +1,140 @@
+body {
+    background: url('https://nancieteresa.files.wordpress.com/2015/01/grunge-travel-background-239421341.jpg');
+    background-size: cover;                   
+    background-repeat:   no-repeat;
+    background-position:  top; 
+}
+
+.sign-up-form {
+    text-align : center;
+    padding-top : 200px
+}
+
+.sign-up-form h1 {
+    font-size: 60px;
+    color: #4e3043;
+    text-align: center;
+    font-family: 'Great Vibes', cursive;
+}
+
+.sign-up-form input[type=text],
+.sign-up-form input[type=password] {
+    /* Size and position */
+    width: 25%;
+    padding: 8px 4px 8px 10px;
+    margin-bottom: 0px;
+    color: black;
+
+    /* Styles */
+    border: 1px solid #4e3043; /* Fallback */
+    border: 1px solid rgba(78,48,67, 0.8);
+    background: whitesmoke;
+    border-radius: 2px;
+    opacity: 75%;
+    box-shadow: 
+        0 1px 0 rgba(255,255,255,0.2), 
+        inset 0 1px 1px rgba(0,0,0,0.1);
+    -webkit-transition: all 0.3s ease-out;
+    -moz-transition: all 0.3s ease-out;
+    -ms-transition: all 0.3s ease-out;
+    -o-transition: all 0.3s ease-out;
+    transition: all 0.3s ease-out;
+
+    /* Font styles */
+    font-family: 'Luckiest Guy', cursive;
+    text-align: center;
+    color: black;
+    font-size: 13px;;
+}
+.sign-up-form input::-webkit-input-placeholder {
+    color: rgba(37,21,26,0.5);
+    text-shadow: 0 1px 0 rgba(255,255,255,0.15);
+}
+
+.sign-up-form input:-moz-placeholder {
+    color: rgba(37,21,26,0.5);
+    text-shadow: 0 1px 0 rgba(255,255,255,0.15);
+    font-family: 'Luckiest Guy';
+}
+
+.sign-up-form input:-ms-input-placeholder {
+    color: rgba(37,21,26,0.5);
+    text-shadow: 0 1px 0 rgba(255,255,255,0.15);
+    font-family: 'Luckiest Guy';
+}
+.sign-up-form input[type=text]:hover,
+.sign-up-form input[type=password]:hover {
+    border-color: #333;
+    font-family: 'Luckiest Guy';
+    font-size: 15px;
+    width: 50%;
+}
+
+.sign-up-form input[type=text]:focus,
+.sign-up-form input[type=password]:focus,
+.sign-up-form input[type=submit]:focus {
+    box-shadow: 
+        0 1px 0 rgba(255,255,255,0.2), 
+        inset 0 1px 1px rgba(0,0,0,0.1),
+        0 0 0 3px rgba(255,255,255,0.15);
+    outline: none;
+    font-family: 'Luckiest Guy';
+    font-size: 15px;
+    width: 50%;
+}
+
+/* Fallback */
+.no-boxshadow .sign-up-form input[type=text]:focus,
+.no-boxshadow .sign-up-form input[type=password]:focus {
+    outline: 1px solid white;
+}
+.sign-up-form input[type=submit] {
+    /* Size and position */
+    width: 25%;
+    padding: 8px 5px;
+    
+    /* Styles */
+    background: linear-gradient(rgba(99, 88, 64, 0.797), rgba(76,49,65,0.7));    
+    border-radius: 5px;
+    border: 1px solid #4e3043;
+    box-shadow: 
+    	inset 0 1px rgba(255,255,255,0.4), 
+    	0 2px 1px rgba(0,0,0,0.1);
+    cursor: pointer;
+    transition: all 0.3s ease-out;
+
+    /* Font styles */
+    color: white;
+    text-shadow: 0 1px 0 rgba(0,0,0,0.3);
+    font-size: 16px;
+    font-weight: bold;
+    font-family: 'Luckiest Guy', cursive;
+}
+
+.sign-up-form input[type=submit]:hover {
+    box-shadow: 
+        inset 0 1px rgba(255,255,255,0.2), 
+        inset 0 20px 30px rgba(99,64,86,0.5);
+    font-weight: bolder;
+    font-size: 23px;
+}
+
+/* Fallback */
+.no-boxshadow .sign-up-form input[type=submit]:hover {
+    background: #594642;
+    font-weight: bolder;
+    font-size: 20px;
+}
+.sign-up-form label {
+    display: none;
+    padding: 0 0 5px 2px;
+    cursor: pointer;
+}
+
+.sign-up-form label:hover ~ input {
+    border-color: #333;
+}
+
+.no-placeholder .sign-up-form label {
+    display: block;
+}
