+++ client/signup.html
@@ -5,7 +5,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
+    <title>Sign-Up</title>
+    <link rel="stylesheet" href="/signup/signup.css">
+    <link href="https://fonts.googleapis.com/css?family=Great+Vibes|Rubik:400i&display=swap" rel="stylesheet">
 </head>
 
 <body>
@@ -23,7 +25,7 @@
             <br />
             <input class="txtbx" id="entrybox" type="password" name="password" placeholder="Create a password..."
                 required><br /><br />
-            <input class="signup-btn" id="signupbutton" type="submit" value="Sign Up!" />
+            <input class="signup-btn" id="signupbutton" type="submit" value="Create your account!" />
         </form>
     </body>
     </form>
